Fix last eslint error
